4-Neural_style_transfer/requirements.txt
tensorflow 
numpy 
matplotlib 
pillow 
gradio